requirements.txt
asgiref==3.7.2
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.2
Django==4.2.6
django-channels==0.7.0
idna==3.4
oauthlib==3.2.2
pdf2image==1.16.3
Pillow==10.1.0
psycopg2-binary==2.9.9
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
PyPDF2==3.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
Wand==0.6.13
